Various Bug Fixes
Fixed a mismatch of ID's, fixed a bug where the contact form was going from uncontrolled to controlled, fixed more drawer issues (finally?)
Adding Loading Spinner to contact form for. while it parses subject line
